Updated to return develops_into and is_a notation.
Working on updating the servlet so that it can drive a jsTree-style interface for SCF.  The servlet now does the sub/superclass computation, and the develops_from graph relationships, and correctly computes the complete "develops_into" relationships from a starting cell type.  This involves following the develops_into relationships from either a term itself, or the subclasses of the term.

The servlet is now outputting text; the next change will be to update it to use JSON.